pyproject.toml
[tool.poetry]
name = "telegram-to-git"
version = "0.1.0"
description = ""
authors = ["Vitor Daniel <vitor036daniel@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_to_git"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.1"
docker = "^6.0.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
gunicorn = "^20.1.0"
python-multipart = "^0.0.5"
pillow = "^9.4.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
